+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.15.0 (?????? ????, ????; functions and variables)
 
 * `$download`, `$openInEditor`, `$select`, `$unwrap` functions for [quick executions](https://github.com/mkloubert/vs-script-commands#quick-execution-)
+* function that do HTTP requests can load headers from files now
 * `$setState` function can also handle async (Promise based) results now
 
 ## 4.14.0 (May 7th, 2017; functions and variables)